requirements.txt
# Dell Switch Port Tracer Web Service - Python Dependencies
# Supports Dell N2000/N3000/N3200 series switches (N2048, N3024P, N3248 models)
# Version: 2.1.1
# Last Updated: August 2025
# Features: Enhanced UI/UX, VLAN Management v2, Database-driven architecture

# Core Web Framework
Flask>=2.3.0

# SSH Connection Library for Dell Switch Communication
paramiko>=3.3.0

# Environment Variable Management
python-dotenv>=1.0.0

# HTTP Basic Authentication Support
Flask-HTTPAuth>=4.7.0

# Database ORM and Migration Support
Flask-SQLAlchemy>=3.0.5
Flask-Migrate>=4.0.5

# PostgreSQL Database Adapter (for Kubernetes deployment)
psycopg2-binary>=2.9.0

# Windows Active Directory Authentication (Production Ready)
ldap3>=2.9.1

# System Monitoring for CPU Safety
psutil>=5.9.0

# Additional Security Dependencies
cryptography>=3.4.8
bcrypt>=3.2.0

# Optional: For production deployment
# gunicorn>=20.1.0
# supervisor>=4.2.4

# Development dependencies (uncomment for development)
# pytest>=7.1.2
# pytest-flask>=1.2.0
# mock>=4.0.3
